Optimize source refresh and ES indexation process

+++ Makefile
@@ -20,7 +20,7 @@ build_front:
 	docker compose -f .docker/frontend/docker-compose.yml build
 	docker compose -f .docker/frontend/docker-compose.yml run --rm front bash -c "npm ci && npx react-scripts build"
 	find ./public -maxdepth 1 -type f -exec rm {} \;
-	rm -r ./public/static ./public/images ./public/locales
+	if [ -e ./public/static ]; then rm -r ./public/static ./public/images ./public/locales; fi
 	mv ./front/build/* ./public/
 	rm -rf ./front/node_modules
